Refactored into Base class for auth and module for s3 handling
